parser: eliminate exception control-flow & remove E.entry
This is preparation for error-recovery which requires replace `loop_handle_undo`
which is hidden in the implementation of `E.entry`.

Solution: inline entry and use `fail` function properly instead of
raise-try-catch control-flow. Also it moves responsibility of driving the parser
from the explanation module `menhir_error_reporting.ml` to `parsing.ml`.