[VSC-18] Fix parseMotokoTyped tests
Problem: The test will fail because the function is not called correctly
and the AST is buried deep inside some nested structures.

Solution: Pass a list of filenames rather than the file contents, get
the AST from `.code[0][0].ast`. `code` contains the result (as opposed
to `diagnostics`), the first `[0]` is where the AST and immediate
imports are (`[1]` has the cache), the second `[0]` has the result of
parsing (since we passed one file name there will be one file here) and
`ast` is the AST.