[VSC-10] Accumulate regions for object fields
Problem: We'd like to track possible origin regions for object fields.
For example, if two objects A and B somehow interacted in the type
checker (from `lub` or `glb` operations), we want to indicate that a
field could come from either A or B or both.

Solution: For `src` and `visibility_src`, add a `srcs` field with the
purpose of tracking those fields. In the typechecker, populate `srcs` as
the typing goes. For `lub`, we compute the unions, and for `glb`, we
compute the intersections. We also perform the union while testing for a
subtyping relation, if it's true.

We also add a way to expose these fields from the compiler using the
S-expression serializer, as well as the types for every node.