[VSC-10] Combine sources when inferring decs
Problem: When we have a declaration like `let a = b`, then `b` might
have an object type with fields, but those fields won't be filled for
`a`. This is due to how `infer_dec` works, since it does nothing useful
with the patterns.

Solution: If we want to combine sources, then call a new function which
will traverse the pattern and type, calling `sub` whenever a `VarP` is
found.

Note: `sub` will "taint" the supertype with the subtype's sources,
adding extraneous fields. It's a good approximation, but not very
accurate.