[VSC-10] Only mutate sources if relation holds
Problem: While we run `rel_fields`, we check `is_rel` and update source
fields if that is the case. However, it's possible that later we'll find
that the relation doesn't hold, but we already "committed" those
changes.

Solution: As suggested by Claudio, have a list that accumulates the
updates in a delayed form using thunks, and only commits them if the
subtyping relation holds.