[VSC-10] Arrange types for nodes used by the LSP
Problem: We arrange types for every node, which can negatively impact
performance. This is because there is no sharing, so there will be a lot
of duplication.

Solution: The solution is a bit hacky, but only arrange types for `DotE`
and the LHS of `DotE` nodes, as those are currently the only ones used
by the language server.