[VSC-10] Tabulate field sources
Problem: Our approach of adding a mutable field to the
`Mo_types.Type.src` type can introduce unwanted mutations in cached
scopes.

Solution: Use a variation of Claudio's suggestion to create a table to
track these sources. This table is mutable to allow it to be easily
passed to `Mo_types.Type` without changing its interface too much. Two
functions are created allowing to convert this table to and from an
immutable map for insertion in scopes. The modules for arranging were
changed to take in this table and fill it while arranging types.